refactor: modularize data loading and chatbot
